Begin implementing per-server configs
This will move most of the bot's config options to be done on a
per-server basis, rather than a global basis. This includes the
following:

- Pterodactyl Panel URL
- Pterodactyl Panel API Key
- Moderator Role
- Admin Role

Admin/Moderator roles will no longer refer to Minecraft directly, since
Pterodactyl can run non-Minecraft game servers (and even non-game
servers), so it doesn't make sense to imply that only Minecraft is
supported.

Server members will also be able to set their own API Key to be used for
the panel as well. This will be limited to those with the Mod and Admin
roles, as otherwise it could cause a conflict should someone not have
permission to interact with another's server. This might be changed in
the future, though.

Additionally, these config options are implemented with slash commands
rather than standard message-parsing commands. Ideally, I'd like to move
everything over to slash commands, but one step at a time!